Fix goto-def on partial methods in LSP
